requirements.txt
lockfile==0.9.1
oursql==0.9.3.1
pysandbox==0.3.5
python-daemon==1.6
sh==1.08
wsgiref==0.1.2
